Fix file upload and serve static files
This commit fixes the file upload functionality in the API by updating the filename generation logic. It now includes the fieldname and the current timestamp. Additionally, it adds a route to serve uploaded files in the API.

In the frontend, a blank line was added in the Home component.

A new package-lock.json file was added in the node_modules directory.